Fix warnings on Visual Studio (#1428)
* Fix warning C4244 in src/layer/normalize.cpp

C4244: '=': conversion from 'double' to 'float', possible loss of data

* Fix warning C4244 in src/layer/requantize.cpp

C4244: 'initializing': conversion from 'double' to 'int', possible loss of data

* Fix warning C4244 in src/mat_pixel_resize.cpp

C4244: '=': conversion from 'double' to 'int', possible loss of data

* Fix warning C4244 in src/mat_pixel.cpp

C4244: '=': conversion from 'int' to 'float', possible loss of data

* Fix warning C4267 in src/modelbin.cpp

C4267: 'initializing': conversion from 'size_t' to 'int', possible loss of data

* Fix warning C4244 in src/layer/batchnorm.cpp

C4244: 'initializing': conversion from 'double' to 'float',
    possible loss of data

* Fix warning C4244 in src/layer/padding.cpp

C4244: 'argument': conversion from 'const float' to 'T', possible loss of data

* Fix warning C4244 in src/layer/priorbox.cpp

C4244: '=': conversion from 'double' to 'float',
    possible loss of data
C4244: 'initializing': conversion from 'double' to 'float',
    possible loss of data

+++ src/layer/requantize.cpp
@@ -29,7 +29,7 @@ Requantize::Requantize()
 
 static inline signed char float2int8(float v)
 {
-    int int32 = round(v);
+    int int32 = static_cast<int>(round(v));
     if (int32 > 127) return 127;
     if (int32 < -127) return -127;
     return (signed char)int32;
